refactor(agents): rename helper modules per review
- agent/chat/chatStatus.ts -> agent/chat/chatUtils.ts
- components/agent/agentChatOperationError.ts -> components/agent/types.ts
