+++ app/src/agent/chat/chatUtils.ts
